util gv_random: operate on 'size_t' values
This pays down some debt from previous changes.

+++ lib/edgepaint/node_distinct_coloring.c
@@ -165,7 +165,7 @@ static void node_distinct_coloring_internal(int scheme, QuadTree qt,
     srand(123);
     iter = -seed;
     for (i = 0; i < iter; i++){
-      seed = gv_random(100000);
+      seed = (int)gv_random(100000);
       node_distinct_coloring_internal2(scheme, qt, weightedQ, A, cdim, accuracy, seed, colors, &color_diff, &color_diff_sum);
       if (color_diff_max < color_diff){
 	seed_max = seed; color_diff_max = color_diff;
